fastapi_backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-jose
passlib[bcrypt]
python-multipart
email-validator
django
python-dotenv
stripe
websockets
wsproto
sendgrid
boto3
